Jwt auth

+++ src/main/java/com/aantivero/wiki/vertx/http/HttpServerVerticle.java
@@ -11,8 +11,13 @@ import io.vertx.core.http.HttpServerOptions;
 import io.vertx.core.json.JsonArray;
 import io.vertx.core.json.JsonObject;
 import io.vertx.core.net.JksOptions;
+import io.vertx.ext.auth.KeyStoreOptions;
+import io.vertx.ext.auth.User;
 import io.vertx.ext.auth.jdbc.JDBCAuth;
+import io.vertx.ext.auth.jwt.JWTAuth;
+import io.vertx.ext.auth.jwt.JWTAuthOptions;
 import io.vertx.ext.jdbc.JDBCClient;
+import io.vertx.ext.jwt.JWTOptions;
 import io.vertx.ext.web.Router;
 import io.vertx.ext.web.RoutingContext;
 import io.vertx.ext.web.client.HttpResponse;
@@ -118,7 +123,49 @@ public class HttpServerVerticle extends AbstractVerticle{
 	    });
 	    //end auth login
 
+	    // jwt token handler on api routes
         Router apiRouter = Router.router(vertx);
+
+	    JWTAuth jwtAuth = JWTAuth.create(vertx,
+		    new JWTAuthOptions()
+			    .setKeyStore(
+			    	new KeyStoreOptions()
+					    .setPath("keystore.jceks")
+					    .setType("jceks")
+					    .setPassword("secret")));
+	    apiRouter.route().handler(JWTAuthHandler.create(jwtAuth, "/api/token"));
+	    // generate jwt tokens
+	    apiRouter.get("/token").handler(context -> {
+	    	JsonObject creds = new JsonObject()
+			    .put("username", context.request().getHeader("login"))
+			    .put("password", context.request().getHeader("password"));
+
+	    	auth.authenticate(creds, authResult -> {
+	    		if (authResult.succeeded()) {
+				    User user = authResult.result();
+				    user.isAuthorized("create", canCreate -> {
+				    	user.isAuthorized("delete", canDelete -> {
+				    		user.isAuthorized("update", canUpdate -> {
+
+				    			String token = jwtAuth.generateToken(new JsonObject()
+								    .put("username", context.request().getHeader("login"))
+								    .put("canCreate", canCreate.succeeded() && canCreate.result())
+								    .put("canDelete", canDelete.succeeded() && canDelete.result())
+								    .put("canUpdate", canUpdate.succeeded() && canUpdate.result()),
+								    new JWTOptions()
+									    .setSubject("Antivero Wiki API")
+								        .setIssuer("Vert.x"));
+
+				    			context.response().putHeader("Content-Type", "text/plain").end(token);
+						    });
+					    });
+				    });
+			    } else {
+	    			context.fail(401);
+			    }
+		    });
+	    });
+
         apiRouter.get("/pages").handler(this::apiRoot);
         apiRouter.get("/pages/:id").handler(this::apiGetPage);
         apiRouter.post().handler(BodyHandler.create());
@@ -155,24 +202,32 @@ public class HttpServerVerticle extends AbstractVerticle{
 	}
 
 	private void apiDeletePage(RoutingContext context) {
-    	int id = Integer.valueOf(context.request().getParam("id"));
+    	if (context.user().principal().getBoolean("canDelete", false)) {
+		    int id = Integer.valueOf(context.request().getParam("id"));
 
-    	dbService.deletePage(id, reply -> {
-    		handleSimpleDbReply(context, reply);
-	    });
+		    dbService.deletePage(id, reply -> {
+			    handleSimpleDbReply(context, reply);
+		    });
+	    } else {
+    		context.fail(401);
+	    }
 	}
 
 	private void apiUpdatePage(RoutingContext context) {
-    	JsonObject page = context.getBodyAsJson();
-    	int id = Integer.valueOf(context.request().getParam("id"));
+    	if (context.user().principal().getBoolean("canUpdate", false)) {
+		    JsonObject page = context.getBodyAsJson();
+		    int id = Integer.valueOf(context.request().getParam("id"));
 
-    	if (!validateJsonPageDocument(context, page, "markdown")) {
-    		return;
+		    if (!validateJsonPageDocument(context, page, "markdown")) {
+			    return;
+		    }
+
+		    dbService.savePage(id, page.getString("markdown"), reply -> {
+			    handleSimpleDbReply(context, reply);
+		    });
+	    } else {
+    		context.fail(401);
 	    }
-
-	    dbService.savePage(id, page.getString("markdown"), reply -> {
-	    	handleSimpleDbReply(context, reply);
-	    });
 	}
 
 	private void handleSimpleDbReply(RoutingContext context, AsyncResult<Void> reply) {
@@ -197,23 +252,27 @@ public class HttpServerVerticle extends AbstractVerticle{
 	}
 
 	private void apiCreate(RoutingContext context) {
-        JsonObject page = context.getBodyAsJson();
-        if (!validateJsonPageDocument(context, page, "name", "markdown")) {
-            return;
-        }
-        dbService.createPage(page.getString("name"), page.getString("markdown"), reply -> {
-            if (reply.succeeded()) {
-                context.response().setStatusCode(201);
-                context.response().putHeader("Content-Type", "application/json");
-                context.response().end(new JsonObject().put("success", true).encode());
-            } else {
-                context.response().setStatusCode(500);
-                context.response().putHeader("Content-Type", "application/json");
-                context.response().end(new JsonObject()
-                    .put("success", false)
-                    .put("error", reply.cause().getMessage()).encode());
-            }
-        });
+    	if (context.user().principal().getBoolean("canCreate",false)) {
+		    JsonObject page = context.getBodyAsJson();
+		    if (!validateJsonPageDocument(context, page, "name", "markdown")) {
+			    return;
+		    }
+		    dbService.createPage(page.getString("name"), page.getString("markdown"), reply -> {
+			    if (reply.succeeded()) {
+				    context.response().setStatusCode(201);
+				    context.response().putHeader("Content-Type", "application/json");
+				    context.response().end(new JsonObject().put("success", true).encode());
+			    } else {
+				    context.response().setStatusCode(500);
+				    context.response().putHeader("Content-Type", "application/json");
+				    context.response().end(new JsonObject()
+					    .put("success", false)
+					    .put("error", reply.cause().getMessage()).encode());
+			    }
+		    });
+	    } else {
+    		context.fail(401);
+	    }
     }
 
     private boolean validateJsonPageDocument(RoutingContext context, JsonObject page, String... expectedKeys) {
